requirements.txt
discord.py==2.3.2
requests==2.31.0
pytz==2024.1
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.42.0
aiohttp==3.9.3